pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cavity-ml"
version = "1.0.0"
description = "Reproducible cavity filter dimension prediction pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "scikit-learn>=1.3",
  "openpyxl>=3.1",
  "pyyaml>=6.0",
  "joblib>=1.3",
  "flask>=3.0",
  "torch>=2.5",
  "matplotlib>=3.8",
  "seaborn>=0.13",
]

[project.optional-dependencies]
test = ["pytest>=8.0"]
dev = ["pytest>=8.0", "ruff>=0.7.0"]

[project.scripts]
prepare-data = "cavity_ml.cli:prepare_data_entry"
validate-data = "cavity_ml.cli:validate_data_entry"
train = "cavity_ml.cli:train_entry"
evaluate = "cavity_ml.cli:evaluate_entry"
predict = "cavity_ml.cli:predict_entry"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
